restante de metodos de clientes ajustados para a Base de dados

+++ Backend-p3/src/main/java/pt/uc/dei/proj3/dto/ClientDto.java
@@ -17,7 +17,7 @@ public class ClientDto {
     // Getters e Setters
     @XmlElement
     public long getId() { return id; }
-    public void setId(int id) { this.id = id; }
+    public void setId(long id) { this.id = id; }
 
     @XmlElement
     public String getNome() { return nome; }
